fix(mcp): break circular ESM imports causing TDZ in tests
Extract shared helpers (resolveProjectDir, failPayload, classifyError, etc.)
from server-handlers.js into shared-helpers.js and error-classifiers.js.
Sub-handlers now import from shared-helpers.js, breaking the circular
server-handlers.js <-> handlers/*.js dependency chain that caused
ReferenceError (TDZ) when Vitest evaluated the module graph.
